Docs and organize upset-react components
