+++ packages/upset/src/components/Header/AttributeScale.tsx
@@ -4,7 +4,7 @@ import { useRecoilValue } from 'recoil';
 import { dimensionsSelector } from '../../atoms/dimensionsAtom';
 import { useScale } from '../../hooks/useScale';
 import translate from '../../utils/transform';
-import { Axis } from '../Axis';
+import { Axis } from '../custom/Axis';
 
 type Props = {
   domain: [number, number];
