+++ packages/upset/src/components/custom/ReactMarkdownWrapper.tsx
@@ -1,6 +1,6 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import Markdown from 'react-markdown';
 
+// Wrapper for the Markdown component to parse and render the alt-txt json
 export default function ReactMarkdownWrapper({ text }: { text: string }) {
   const components = {
     // eslint-disable-next-line react/no-unstable-nested-components, jsx-a11y/heading-has-content
