+++ packages/upset/src/utils/exports.ts
@@ -3,6 +3,12 @@ import {
 } from '@visdesignlab/upset2-core';
 import { UpsetProvenance } from '../provenance';
 
+/**
+ * Retrieves accessible data from rows.
+ * @param rows - The rows containing the data.
+ * @param includeId - Optional parameter to include the ID in the accessible data.
+ * @returns The accessible data.
+ */
 export const getAccessibleData = (rows: Rows, includeId = false): AccessibleData => {
   const data = { values: {} } as AccessibleData;
   Object.values(rows.values).forEach((r: Row) => {
@@ -33,6 +39,12 @@ export const getAccessibleData = (rows: Rows, includeId = false): AccessibleData
   return data;
 };
 
+/**
+ * Downloads a JSON file with the given filename and content.
+ *
+ * @param filename - The name of the file to be downloaded.
+ * @param json - The JSON content to be downloaded.
+ */
 const downloadJSON = (filename: string, json: string): void => {
   const blob = new Blob([json], { type: 'application/json' });
   const href = URL.createObjectURL(blob);
@@ -47,6 +59,11 @@ const downloadJSON = (filename: string, json: string): void => {
   URL.revokeObjectURL(href);
 };
 
+/**
+ * Generates element names based on the provided rows.
+ * @param rows - The rows containing element names.
+ * @returns The updated rows with modified element names.
+ */
 const generateElementName = (rows: Rows): Rows => {
   const newRows = { ...rows };
 
@@ -78,6 +95,13 @@ const generateElementName = (rows: Rows): Rows => {
   return newRows;
 };
 
+/**
+ * Retrieves the alternative text grammar config based on the provided state, data, and rows.
+ * @param state - The UpsetConfig state object.
+ * @param data - The CoreUpsetData object.
+ * @param rows - The Rows object.
+ * @returns The AltTextConfig object.
+ */
 export const getAltTextConfig = (state: UpsetConfig, data: CoreUpsetData, rows: Rows): AltTextConfig => {
   let dataObj = state as AltTextConfig;
 
@@ -93,6 +117,12 @@ export const getAltTextConfig = (state: UpsetConfig, data: CoreUpsetData, rows: 
   return dataObj;
 };
 
+/**
+ * Exports the state of the Upset plot.
+ * @param provenance - The UpsetProvenance object that tracks the state changes of the Upset component.
+ * @param data - Optional. The CoreUpsetData data of the Upset component. (Raw Data)
+ * @param rows - Optional. The rows data of the Upset component. (Processed Data)
+ */
 export const exportState = (provenance: UpsetProvenance, data?: CoreUpsetData, rows?: Rows): void => {
   let filename = `upset_state_${new Date().toJSON().slice(0, 10)}`;
   let dataObj = provenance.getState() as UpsetConfig & { rawData?: CoreUpsetData; processedData?: Rows; accessibleProcessedData?: AccessibleData };
@@ -113,12 +143,21 @@ export const exportState = (provenance: UpsetProvenance, data?: CoreUpsetData, r
   downloadJSON(filename, json);
 };
 
+/**
+ * Exports the raw data as a JSON file.
+ * @param data - The core upset data to be exported.
+ */
 export const exportRawData = (data: CoreUpsetData): void => {
   const filename = `upset_data_raw_${new Date().toJSON().slice(0, 10)}`;
   const json = JSON.stringify(data, null, 2);
   downloadJSON(filename, json);
 };
 
+/**
+ * Exports processed data as a JSON file.
+ * @param rows - The data to be exported.
+ * @param accessible - Optional parameter indicating whether the exported data should fit the accessible datatype.
+ */
 export const exportProcessedData = (rows: Rows, accessible?: boolean): void => {
   const filename = `upset_data_${new Date().toJSON().slice(0, 10)}`;
   const data: Rows | AccessibleData = (accessible) ? getAccessibleData(rows) : rows;
